+++ crypto/sshkeys/private_key.go
@@ -2,12 +2,13 @@ package sshkeys
 
 import (
 	"encoding/pem"
+	"strings"
 
+	"github.com/ScaleFT/sshkeys"
 	"golang.org/x/crypto/ssh"
-)
+	"gopkg.in/errgo.v1"
 
-var (
-	implementedCiphers = []string{"DES-EDE3-CBC", "AES-128-CBC"}
+	"github.com/Scalingo/cli/term"
 )
 
 type PrivateKey struct {
@@ -16,27 +17,37 @@ type PrivateKey struct {
 	PasswordMethod
 }
 
+type PasswordMethod func(prompt string) (string, error)
+
 func (p *PrivateKey) Signer() (ssh.Signer, error) {
+	if p.IsEncrypted() {
+		if p.PasswordMethod == nil {
+			p.PasswordMethod = term.Password
+		}
+
+		password, err := p.PasswordMethod("Encrypted SSH Key, password: ")
+		if err != nil {
+			return nil, errgo.Notef(err, "fail to get the SSH key password")
+		}
+
+		return sshkeys.ParseEncryptedPrivateKey(pem.EncodeToMemory(p.Block), []byte(password))
+	}
+
 	return ssh.ParsePrivateKey(pem.EncodeToMemory(p.Block))
 }
 
 func (p *PrivateKey) IsEncrypted() bool {
-	return p.Block.Headers["Proc-Type"] == "4,ENCRYPTED"
+	return p.Block.Headers["Proc-Type"] == "4,ENCRYPTED" || p.isOpenSSHFormatEncrypted()
 }
 
-func (p *PrivateKey) IsCipherImplemented(cipher string) bool {
-	for _, c := range implementedCiphers {
-		if c == cipher {
-			return true
-		}
+func (p *PrivateKey) isOpenSSHFormatEncrypted() bool {
+	if p.Block.Type != "OPENSSH PRIVATE KEY" {
+		return false
+	}
+
+	_, err := ssh.ParseRawPrivateKey(pem.EncodeToMemory(p.Block))
+	if err != nil {
+		return strings.Contains(err.Error(), "cannot decode encrypted private keys")
 	}
 	return false
 }
-
-func DummyPasswordMethod(password string) PasswordMethod {
-	return func(prompt string) (string, error) {
-		return prompt, nil
-	}
-}
-
-type PasswordMethod func(prompt string) (string, error)